perform loss computations in fp32

+++ skyrl-tx/tx/tinker/engine.py
@@ -182,7 +182,7 @@ class TinkerEngine:
             model = nnx.merge(self.graphdef, lora_params, non_lora_params)
             logits = model(input_ids, attention_mask=attention_mask, adapter_indices=adapter_indices)[
                 "logits"
-            ]  # [B, T, V]
+            ].astype(jnp.float32)  # [B, T, V]
             logprobs = jax.nn.log_softmax(logits, axis=-1)  # [B, T, V]
             target_logprobs = jnp.take_along_axis(logprobs, target_ids[..., None], axis=-1).squeeze(-1)
